refactor: split svg renderer
